requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
exceptiongroup==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.1
PyYAML==6.0
-e git+https://gitlab.com/wowx-baas/ewallet-system/functional/common-libraries/repository-pattern.git@9029a6af18c7d78de2ae1d37cd2e7429e70d993a#egg=repository_pattern
requests==2.31.0
SQLAlchemy==1.4.48
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
